# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export type ApiErrorErrorCode = typeof ApiErrorErrorCode[keyof typeof ApiErrorErrorCode];


export const ApiErrorErrorCode = {
  bad_request: 'bad_request',
  unauthorized: 'unauthorized',
  forbidden: 'forbidden',
  not_found: 'not_found',
  payload_too_large: 'payload_too_large',
  too_many_requests: 'too_many_requests',
  internal_error: 'internal_error',
} as const;

export type ApiErrorError = {
  code: ApiErrorErrorCode;
  message: string;
};

export interface ApiError {
  error: ApiErrorError;
}

export type MembershipRole = typeof MembershipRole[keyof typeof MembershipRole];


export const MembershipRole = {
  OWNER: 'OWNER',
  ADMIN: 'ADMIN',
  MEMBER: 'MEMBER',
  VIEWER: 'VIEWER',
} as const;

export type OrganizationPlan = typeof OrganizationPlan[keyof typeof OrganizationPlan];


export const OrganizationPlan = {
  STARTER: 'STARTER',
  PROFESSIONAL: 'PROFESSIONAL',
  ENTERPRISE: 'ENTERPRISE',
} as const;

export type Capability = typeof Capability[keyof typeof Capability];


export const Capability = {
  workspacemanage: 'workspace.manage',
  workspacetransfer: 'workspace.transfer',
  membersmanage: 'members.manage',
  clientswrite: 'clients.write',
  scanswrite: 'scans.write',
  tasksmanage: 'tasks.manage',
  reportsgenerate: 'reports.generate',
  reportsview: 'reports.view',
  auditviewAll: 'audit.viewAll',
  auditviewOwn: 'audit.viewOwn',
} as const;

export interface LoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface SwitchOrganizationInput {
  /** @minLength 1 */
  organizationId: string;
}

export interface SessionAccount {
  id: string;
  name: string;
  email: string;
  /** Nullable in the database; the server derives it from the name so the contract always carries a value. */
  avatarInitials: string;
  preferredLocale: string;
}

export interface SessionOrganization {
  id: string;
  name: string;
  plan: OrganizationPlan;
}

export interface SessionMembership {
  organizationId: string;
  organizationName: string;
  role: MembershipRole;
}

/**
 * Invariant: organization.id always appears in memberships, role is always that membership's role, and capabilities is always the derivation of role through the server's permission matrix.
 */
export interface Session {
  user: SessionAccount;
  organization: SessionOrganization;
  role: MembershipRole;
  capabilities: Capability[];
  memberships: SessionMembership[];
}

export interface HealthStatus {
  status: string;
}

export type PeppolStatus = typeof PeppolStatus[keyof typeof PeppolStatus];


export const PeppolStatus = {
  READY: 'READY',
  CONFIGURING: 'CONFIGURING',
  AT_RISK: 'AT_RISK',
  NOT_REGISTERED: 'NOT_REGISTERED',
} as const;

export type RiskSeverity = typeof RiskSeverity[keyof typeof RiskSeverity];


export const RiskSeverity = {
  info: 'info',
  warning: 'warning',
  critical: 'critical',
} as const;

export interface ReadinessAssessmentInput {
  participantRegistered: boolean;
  receivingAddressConfigured: boolean;
  peppolCapableSoftware: boolean;
  certificateValid: boolean;
  successfulTestInvoice: boolean;
  checkedAt?: string;
  /** @minLength 1 */
  source?: string;
}

export interface ReadinessFactor {
  key: string;
  label: string;
  weight: number;
  passed: boolean;
  earnedPoints: number;
  message: string;
}

export interface RiskIndicator {
  code: string;
  label: string;
  severity: RiskSeverity;
  message: string;
  remediation: string;
}

export interface ReadinessAssessment {
  companyId: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  score: number;
  status: PeppolStatus;
  riskLevel: RiskSeverity;
  factors: ReadinessFactor[];
  risks: RiskIndicator[];
  calculatedAt: string;
}

export interface DashboardKpis {
  totalCompanies: number;
  peppolReady: number;
  actionRequired: number;
  highRisk: number;
  notRegistered: number;
  averageScore: number;
}

export interface ReadinessBreakdown {
  ready: number;
  configuring: number;
  atRisk: number;
  notRegistered: number;
}

export interface ReadinessTrendPoint {
  period: string;
  label: string;
  averageScore: number;
}

export interface DashboardAction {
  code: string;
  label: string;
  count: number;
  severity: RiskSeverity;
}

export interface DashboardIncident {
  id: string;
  title: string;
  /** @nullable */
  companyId: string | null;
  /** @nullable */
  companyName: string | null;
  severity: RiskSeverity;
  status: string;
  occurredAt: string;
}

export interface DashboardCompany {
  id: string;
  name: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  accountingPackage: string | null;
  status: PeppolStatus;
  score: number;
  /** @nullable */
  lastCheckedAt: string | null;
  risks: RiskIndicator[];
}

export type ReadinessDashboardOrganization = {
  id: string;
  name: string;
};

export interface ReadinessDashboard {
  organization: ReadinessDashboardOrganization;
  generatedAt: string;
  kpis: DashboardKpis;
  breakdown: ReadinessBreakdown;
  trend: ReadinessTrendPoint[];
  actions: DashboardAction[];
  incidents: DashboardIncident[];
  companies: DashboardCompany[];
}

/**
 * The request body is missing, malformed, or fails validation
 */
export type BadRequestResponse = ApiError;

/**
 * The request body exceeds the size limit
 */
export type PayloadTooLargeResponse = ApiError;

/**
 * Too many attempts; retry after the number of seconds in Retry-After
 */
export type TooManyRequestsResponse = ApiError;