# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: readiness
    description: Peppol readiness scoring and dashboard operations
  - name: auth
    description: Authentication, active organization, and session identity
paths:
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in
      description: >-
        Verifies credentials and starts a session. Accepts JSON only; form
        bodies are rejected so a cross-site form cannot sign a visitor in.
        Attempts are limited per client address and per email address.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Sign out
      description: Destroys the current session.
      responses:
        "204":
          description: Session destroyed
  /auth/session:
    get:
      operationId: getSession
      tags: [auth]
      summary: Get the current session
      description: Returns the signed-in user, active organization, role, capabilities, and memberships.
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/organization:
    post:
      operationId: switchOrganization
      tags: [auth]
      summary: Switch the active organization
      description: Moves the session to another organization the signed-in user is a member of.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwitchOrganizationInput"
      responses:
        "200":
          description: Session moved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Not a member of that organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /readiness/dashboard:
    get:
      operationId: getReadinessDashboard
      tags: [readiness]
      summary: Get readiness dashboard data
      description: >-
        Returns readiness KPIs, trends, risks, incidents, and companies for the
        organization the session is acting in. The tenant is derived from the
        session and is never accepted as a parameter.
      responses:
        "200":
          description: Readiness dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessDashboard"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /companies/{companyId}/readiness/calculate:
    post:
      operationId: calculateCompanyReadiness
      tags: [readiness]
      summary: Calculate company readiness
      description: >-
        Calculates and persists a new readiness assessment from weighted Peppol
        criteria. The company must belong to the session's organization.
      parameters:
        - in: path
          name: companyId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReadinessAssessmentInput"
      responses:
        "200":
          description: Calculated readiness assessment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessAssessment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Not permitted for this role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Company not found in this organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
components:
  responses:
    BadRequest:
      description: The request body is missing, malformed, or fails validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    PayloadTooLarge:
      description: The request body exceeds the size limit
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    TooManyRequests:
      description: Too many attempts; retry after the number of seconds in Retry-After
      headers:
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: number
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
    ApiError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                [
                  bad_request,
                  unauthorized,
                  forbidden,
                  not_found,
                  payload_too_large,
                  too_many_requests,
                  internal_error,
                ]
            message:
              type: string
          required: [code, message]
      required: [error]
    MembershipRole:
      type: string
      enum: [OWNER, ADMIN, MEMBER, VIEWER]
    OrganizationPlan:
      type: string
      enum: [STARTER, PROFESSIONAL, ENTERPRISE]
    Capability:
      type: string
      enum:
        - workspace.manage
        - workspace.transfer
        - members.manage
        - clients.write
        - scans.write
        - tasks.manage
        - reports.generate
        - reports.view
        - audit.viewAll
        - audit.viewOwn
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 1
      required: [email, password]
    SwitchOrganizationInput:
      type: object
      properties:
        organizationId:
          type: string
          minLength: 1
      required: [organizationId]
    SessionAccount:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        avatarInitials:
          type: string
          description: Nullable in the database; the server derives it from the name so the contract always carries a value.
        preferredLocale:
          type: string
      required: [id, name, email, avatarInitials, preferredLocale]
    SessionOrganization:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        plan:
          $ref: "#/components/schemas/OrganizationPlan"
      required: [id, name, plan]
    SessionMembership:
      type: object
      properties:
        organizationId:
          type: string
        organizationName:
          type: string
        role:
          $ref: "#/components/schemas/MembershipRole"
      required: [organizationId, organizationName, role]
    Session:
      type: object
      description: >-
        Invariant: organization.id always appears in memberships, role is
        always that membership's role, and capabilities is always the
        derivation of role through the server's permission matrix.
      properties:
        user:
          $ref: "#/components/schemas/SessionAccount"
        organization:
          $ref: "#/components/schemas/SessionOrganization"
        role:
          $ref: "#/components/schemas/MembershipRole"
        capabilities:
          type: array
          items:
            $ref: "#/components/schemas/Capability"
        memberships:
          type: array
          items:
            $ref: "#/components/schemas/SessionMembership"
      required: [user, organization, role, capabilities, memberships]
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    PeppolStatus:
      type: string
      enum: [READY, CONFIGURING, AT_RISK, NOT_REGISTERED]
    RiskSeverity:
      type: string
      enum: [info, warning, critical]
    ReadinessAssessmentInput:
      type: object
      properties:
        participantRegistered:
          type: boolean
        receivingAddressConfigured:
          type: boolean
        peppolCapableSoftware:
          type: boolean
        certificateValid:
          type: boolean
        successfulTestInvoice:
          type: boolean
        checkedAt:
          type: string
          format: date-time
        source:
          type: string
          minLength: 1
      required:
        - participantRegistered
        - receivingAddressConfigured
        - peppolCapableSoftware
        - certificateValid
        - successfulTestInvoice
    ReadinessFactor:
      type: object
      properties:
        key:
          type: string
        label:
          type: string
        weight:
          type: number
        passed:
          type: boolean
        earnedPoints:
          type: number
        message:
          type: string
      required: [key, label, weight, passed, earnedPoints, message]
    RiskIndicator:
      type: object
      properties:
        code:
          type: string
        label:
          type: string
        severity:
          $ref: "#/components/schemas/RiskSeverity"
        message:
          type: string
        remediation:
          type: string
      required: [code, label, severity, message, remediation]
    ReadinessAssessment:
      type: object
      properties:
        companyId:
          type: string
        score:
          type: number
          minimum: 0
          maximum: 100
        status:
          $ref: "#/components/schemas/PeppolStatus"
        riskLevel:
          $ref: "#/components/schemas/RiskSeverity"
        factors:
          type: array
          items:
            $ref: "#/components/schemas/ReadinessFactor"
        risks:
          type: array
          items:
            $ref: "#/components/schemas/RiskIndicator"
        calculatedAt:
          type: string
          format: date-time
      required:
        [companyId, score, status, riskLevel, factors, risks, calculatedAt]
    DashboardKpis:
      type: object
      properties:
        totalCompanies:
          type: number
        peppolReady:
          type: number
        actionRequired:
          type: number
        highRisk:
          type: number
        notRegistered:
          type: number
        averageScore:
          type: number
      required:
        [totalCompanies, peppolReady, actionRequired, highRisk, notRegistered, averageScore]
    ReadinessBreakdown:
      type: object
      properties:
        ready:
          type: number
        configuring:
          type: number
        atRisk:
          type: number
        notRegistered:
          type: number
      required: [ready, configuring, atRisk, notRegistered]
    ReadinessTrendPoint:
      type: object
      properties:
        period:
          type: string
        label:
          type: string
        averageScore:
          type: number
      required: [period, label, averageScore]
    DashboardAction:
      type: object
      properties:
        code:
          type: string
        label:
          type: string
        count:
          type: number
        severity:
          $ref: "#/components/schemas/RiskSeverity"
      required: [code, label, count, severity]
    DashboardIncident:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        companyId:
          type: [string, "null"]
        companyName:
          type: [string, "null"]
        severity:
          $ref: "#/components/schemas/RiskSeverity"
        status:
          type: string
        occurredAt:
          type: string
          format: date-time
      required: [id, title, companyId, companyName, severity, status, occurredAt]
    DashboardCompany:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: [string, "null"]
        accountingPackage:
          type: [string, "null"]
        status:
          $ref: "#/components/schemas/PeppolStatus"
        score:
          type: number
        lastCheckedAt:
          type: [string, "null"]
          format: date-time
        risks:
          type: array
          items:
            $ref: "#/components/schemas/RiskIndicator"
      required: [id, name, email, accountingPackage, status, score, lastCheckedAt, risks]
    ReadinessDashboard:
      type: object
      properties:
        organization:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
          required: [id, name]
        generatedAt:
          type: string
          format: date-time
        kpis:
          $ref: "#/components/schemas/DashboardKpis"
        breakdown:
          $ref: "#/components/schemas/ReadinessBreakdown"
        trend:
          type: array
          items:
            $ref: "#/components/schemas/ReadinessTrendPoint"
        actions:
          type: array
          items:
            $ref: "#/components/schemas/DashboardAction"
        incidents:
          type: array
          items:
            $ref: "#/components/schemas/DashboardIncident"
        companies:
          type: array
          items:
            $ref: "#/components/schemas/DashboardCompany"
      required:
        [organization, generatedAt, kpis, breakdown, trend, actions, incidents, companies]